dedupe log cleanup day normalization
